+++ src/changeBoardQuickStep.ts
@@ -10,7 +10,7 @@ export async function changeBoardQuickStep(context: ExtensionContext, project: Z
     let boards: ZephyrBoard[] = [];
     await vscode.window.withProgress({
       location: vscode.ProgressLocation.Notification,
-      title: "Updating available boards list",
+      title: "Updating available boards list. It might take a while...",
       cancellable: false,
       }, async () => {
         boards = await getSupportedBoards(westWorkspace, project);